+++ lib/statistics/tasks.rb
@@ -5,3 +5,4 @@ end
 
 require_relative 'tasks/connpass'
 require_relative 'tasks/doorkeeper'
+require_relative 'tasks/facebook'
